+++ CHANGELOG.md
@@ -5,7 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.3] - 2023-07-18 :cat:
+
+- Adds support for icons in cards (by @Andy2003).
+- Adds support for anchors with target="_blank" in cards.
+
 ## [1.0.2] - 2023-04-25
+
 - Fixes detail in the `contribs` plugin: when the name is specified for a
   contributor by email address, it is used.
 - Improves `pyproject.toml`.